+++ lib/flutter_waya.dart
@@ -3,8 +3,8 @@ library flutter_waya;
 export 'package:crypto/crypto.dart';
 export 'package:dio/dio.dart';
 export 'package:flutter_easyrefresh/easy_refresh.dart';
-export 'package:http_parser/http_parser.dart';
 export 'package:flutter_localizations/flutter_localizations.dart';
+export 'package:http_parser/http_parser.dart';
 
 export 'constant/constant.dart';
 export 'extension/extension.dart';